+++ huaweicloud-sdk-projectman/huaweicloudsdkprojectman/v4/region/projectman_region.py
@@ -9,13 +9,13 @@ class ProjectManRegion:
     def __init__(self):
         pass
 
-    CN_NORTH_4 = Region(id="cn-north-4", endpoint="https://projectman.cn-north-4.myhuaweicloud.com")
+    CN_NORTH_4 = Region(id="cn-north-4", endpoint="https://projectman-ext.cn-north-4.myhuaweicloud.com")
 
-    CN_NORTH_1 = Region(id="cn-north-1", endpoint="https://projectman.cn-north-1.myhuaweicloud.com")
+    CN_NORTH_1 = Region(id="cn-north-1", endpoint="https://projectman-ext.cn-north-1.myhuaweicloud.com")
 
-    CN_EAST_2 = Region(id="cn-east-2", endpoint="https://projectman.cn-east-2.myhuaweicloud.com")
+    CN_EAST_2 = Region(id="cn-east-2", endpoint="https://projectman-ext.cn-east-2.myhuaweicloud.com")
 
-    CN_SOUTH_1 = Region(id="cn-south-1", endpoint="https://projectman.cn-south-1.myhuaweicloud.com")
+    CN_SOUTH_1 = Region(id="cn-south-1", endpoint="https://projectman-ext.cn-south-1.myhuaweicloud.com")
 
     static_fields = types.MappingProxyType({
         "cn-north-4": CN_NORTH_4,